+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Fix hidden click fallback in `web.click`.
 * Fix `web.type` not clearing the previous value.
 * `web.selectWindow` will automatically wait for window to appear when using `title` locator.
+* Add optional `notOlderThan` argument to `twilio.getLastSms`.
 
 ## v0.32.0 (2018-08-09)
 * Fix `web.isAlertPresent`.